feat: :bug: fix next auth. split option from route for use to call api with token
