+++ package.json
@@ -16,6 +16,7 @@
     "@mui/material": "^5.16.7",
     "@mui/x-date-pickers": "^7.14.0",
     "@reduxjs/toolkit": "^2.2.8",
+    "axios": "^1.7.7",
     "dayjs": "^1.11.13",
     "next": "14.2.5",
     "next-auth": "^4.24.10",
