+++ src/interfaces/next-auth.d.ts
@@ -1,0 +1,13 @@
+import NextAuth from 'next-auth'
+
+declare module "next-auth" {
+    interface Session {
+        user: {
+            _id: string,
+            name: string,
+            email: string,
+            role: string,
+            token: string
+        }
+    }
+}
